rectifications after review
