+++ server/src/router.ts
@@ -8,14 +8,11 @@ const router = express.Router();
 
 // Define item-related routes
 import itemActions from "./modules/item/itemActions";
-
 router.get("/api/items", itemActions.browse);
 router.get("/api/items/:id", itemActions.read);
 router.post("/api/items", itemActions.add);
 
-// artist-related routes
 import artistActions from "./modules/artist/artistActions";
-
 router.get("/api/artist/:id", artistActions.read);
 /* ************************************************************************* */
